+++ libs/vm-modules/include/vm_modules/ml/model/model.hpp
@@ -31,6 +31,12 @@ class Module;
 }
 
 namespace ml {
+
+namespace dataloaders {
+template <typename TensorType>
+class TensorDataLoader;
+}  // namespace dataloaders
+
 namespace model {
 
 template <typename TensorType>
