+++ libs/vm-modules/src/ml/model/model.cpp
@@ -16,14 +16,17 @@
 //
 //------------------------------------------------------------------------------
 
-#include "vm_modules/ml/model/model.hpp"
-
+#include "core/byte_array/decoders.hpp"
 #include "core/serializers/counter.hpp"
+#include "ml/dataloaders/tensor_dataloader.hpp"
 #include "ml/layers/fully_connected.hpp"
 #include "ml/model/sequential.hpp"
 #include "ml/ops/loss_functions/mean_square_error_loss.hpp"
 #include "ml/ops/loss_functions/types.hpp"
+#include "ml/ops/metrics/types.hpp"
+#include "ml/serializers/ml_types.hpp"
 #include "vm/module.hpp"
+#include "vm_modules/ml/model/model.hpp"
 #include "vm_modules/ml/model/model_estimator.hpp"
 #include "vm_modules/use_estimator.hpp"
